feat(expasion): now all is in the norm and working
