+++ config.mk
@@ -28,12 +28,16 @@ SRC_FILES = main.c \
 			$(AST_DIR)/parse_cmd_rules.c \
 			$(EVAL_DIR)/eval.c \
 			$(EVAL_DIR)/eval_nodes.c \
+			$(EVAL_DIR)/eval_subshell.c \
 			$(EVAL_DIR)/eval_helpers.c \
 			$(EVAL_DIR)/eval_util.c \
 			$(EVAL_DIR)/cmd_res.c \
 			$(EVAL_DIR)/util.c \
 			$(EVAL_DIR)/redir/redir.c \
+			$(EVAL_DIR)/redir/redir_utils.c \
 			$(EVAL_DIR)/redir/here_doc.c \
+			$(EVAL_DIR)/redir/here_doc_utils.c \
+			$(EVAL_DIR)/redir/here_doc_cleanup.c \
 			$(EVAL_DIR)/redir/here_doc_write.c \
 			$(EVAL_DIR)/redir/here_doc_write_util.c \
 			$(EVAL_DIR)/pipe/child_parent.c \
@@ -48,12 +52,19 @@ SRC_FILES = main.c \
 			$(EVAL_DIR)/expand/expand_quotes.c \
 			$(EVAL_DIR)/expand/expand_wildcard.c \
 			$(EVAL_DIR)/expand/expand_wildcard_utils.c \
+			$(EVAL_DIR)/expand/expand_wildcard_list.c \
+			$(EVAL_DIR)/expand/expand_ctx.c \
+			$(EVAL_DIR)/expand/expand_str_utils.c \
+			$(EVAL_DIR)/expand/expand_args_utils.c \
+			$(EVAL_DIR)/expand/expand_special_vars.c \
+			$(EVAL_DIR)/expand/expand_utils.c \
 			$(EVAL_DIR)/builtin/cd.c \
 			$(EVAL_DIR)/builtin/echo.c \
 			$(EVAL_DIR)/builtin/env.c \
 			$(EVAL_DIR)/builtin/env_util.c \
 			$(EVAL_DIR)/builtin/exit.c \
 			$(EVAL_DIR)/builtin/export.c \
+			$(EVAL_DIR)/builtin/export_helpers.c \
 			$(EVAL_DIR)/builtin/export_args_util.c \
 			$(EVAL_DIR)/builtin/export_no_args_util.c \
 			$(EVAL_DIR)/builtin/pwd.c \
