+++ src/eval/cmd/cmd_str_helpers.c
@@ -11,8 +11,41 @@
 /* ************************************************************************** */
 
 #include "eval.h"
+#include "ast.h"
 #include "libft.h"
 #include <unistd.h>
+#include <stdlib.h>
+
+int	count_cmd_suffix(t_ast *cmd_suffix)
+{
+	int		count;
+	t_ast	*node;
+
+	count = 0;
+	node = cmd_suffix;
+	while (node)
+	{
+		if (node->u_ast.s_cmd_suffix.word)
+			count++;
+		node = node->u_ast.s_cmd_suffix.cmd_suffix;
+	}
+	return (count);
+}
+
+void	free_cmd_str(char **cmd_str)
+{
+	int	i;
+
+	if (!cmd_str)
+		return ;
+	i = 0;
+	while (cmd_str[i])
+	{
+		free(cmd_str[i]);
+		i++;
+	}
+	free(cmd_str);
+}
 
 char	**get_bin_paths(t_shell_env *env)
 {
@@ -40,4 +73,24 @@ int	check_path(char *path, char *argv)
 	result = access(full_path, X_OK);
 	free(full_path);
 	return (result == 0);
+}
+
+char	*resolve_command_path(char *cmd_name, t_shell_env *env)
+{
+	char	**bin_paths;
+	int		i;
+	char	*found_path;
+
+	if (ft_strchr(cmd_name, '/') || check_builtin(cmd_name))
+		return (ft_strdup(""));
+	bin_paths = get_bin_paths(env);
+	i = 0;
+	while (bin_paths[i] && !check_path(bin_paths[i], cmd_name))
+		i++;
+	if (bin_paths[i])
+		found_path = ft_strdup(bin_paths[i]);
+	else
+		found_path = ft_strdup("");
+	free_cmd_str(bin_paths);
+	return (found_path);
 }
