+++ src/eval/eval_nodes.c
@@ -12,6 +12,7 @@
 
 #include "ast.h"
 #include "eval.h"
+#include "libft.h"
 #include "minishell_signal.h"
 #include <stdlib.h>
 #include <sys/wait.h>
@@ -41,52 +42,7 @@ t_cmd_response	*eval_pipe(t_ast *shell_ast, t_shell_env *env)
 	return (res);
 }
 
-static void	subshell_child(t_ast *shell_ast, t_shell_env *env)
-{
-	t_shell_response	*sub_res;
-	int					code;
-
-	setup_fork_signal(0);
-	if (env)
-		env->interactive_owner = 0;
-	if (eval_io_file(shell_ast->u_ast.s_subshell.io_file, env) == -1)
-		child_exit(env, shell_ast, 1);
-	sub_res = eval_ast(shell_ast->u_ast.s_subshell.and_or, env);
-	code = 0;
-	if (sub_res)
-		code = sub_res->exit_code;
-	if (sub_res)
-		free(sub_res);
-	child_exit(env, shell_ast, code);
-}
-
-t_cmd_response	*eval_subshell(t_ast *shell_ast, t_shell_env *env)
-{
-	pid_t			pid;
-	int				status;
-	t_cmd_response	*cmd_res;
-
-	if (!shell_ast)
-		return (NULL);
-	pid = fork();
-	if (pid == -1)
-		return (NULL);
-	if (pid == 0)
-		subshell_child(shell_ast, env);
-	setup_fork_signal(pid);
-	waitpid(pid, &status, 0);
-	setup_nonfork_signal();
-	cmd_res = create_cmd_res();
-	if (!cmd_res)
-		return (NULL);
-	if (WIFEXITED(status))
-		cmd_res->exit_code = WEXITSTATUS(status);
-	else if (WIFSIGNALED(status))
-		cmd_res->exit_code = 128 + WTERMSIG(status);
-	return (cmd_res);
-}
-
-static t_cmd_response	*eval_node(t_ast *node, t_shell_env *env)
+t_cmd_response	*eval_node(t_ast *node, t_shell_env *env)
 {
 	if (node->type == AST_SIMPLE_CMD)
 		return (eval_cmd(node, env));
@@ -127,16 +83,30 @@ t_cmd_response	*eval_and_or(t_ast *shell_ast, t_shell_env *env)
 	return (left_res);
 }
 
-void	eval_subshell_in_pipe(t_ast *shell_ast, t_shell_env *env)
+t_shell_response	*eval_ast(t_ast *shell_ast, t_shell_env *env)
 {
-	t_shell_response	*sub_res;
-	int					code;
+	t_shell_response	*res;
+	t_cmd_response		*cmd_res;
 
-	sub_res = eval_ast(shell_ast->u_ast.s_subshell.and_or, env);
-	code = 0;
-	if (sub_res)
-		code = sub_res->exit_code;
-	if (sub_res)
-		free(sub_res);
-	child_exit(env, shell_ast, code);
+	if (!shell_ast)
+		return (NULL);
+	if (process_heredocs(shell_ast, env) == -1)
+	{
+		cleanup_heredoc_files(shell_ast);
+		if (env)
+			env->last_exit_code = 130;
+		return (NULL);
+	}
+	cmd_res = eval_node(shell_ast, env);
+	cleanup_heredoc_files(shell_ast);
+	if (!cmd_res)
+		return (NULL);
+	res = ft_calloc(1, sizeof(t_shell_response));
+	if (!res)
+		return (destroy_cmd_res(cmd_res));
+	res->exit_code = cmd_res->exit_code;
+	if (env)
+		env->last_exit_code = res->exit_code;
+	free(cmd_res);
+	return (res);
 }
