+++ src/eval/redir/here_doc.c
@@ -44,22 +44,23 @@ static int	process_simple_cmd_heredocs(t_ast *node, t_shell_env *env)
 	return (0);
 }
 
+static int	process_binary_heredoc(t_ast *left, t_ast *right, t_shell_env *env)
+{
+	if (process_heredocs(left, env) == -1)
+		return (-1);
+	return (process_heredocs(right, env));
+}
+
 int	process_heredocs(t_ast *node, t_shell_env *env)
 {
 	if (!node)
 		return (0);
 	if (node->type == AST_PIPE_SEQ)
-	{
-		if (process_heredocs(node->u_ast.s_pipe_seq.left, env) == -1)
-			return (-1);
-		return (process_heredocs(node->u_ast.s_pipe_seq.right, env));
-	}
+		return (process_binary_heredoc(node->u_ast.s_pipe_seq.left,
+				node->u_ast.s_pipe_seq.right, env));
 	if (node->type == AST_AND_OR || node->type == AST_LIST)
-	{
-		if (process_heredocs(node->u_ast.s_and_or.left, env) == -1)
-			return (-1);
-		return (process_heredocs(node->u_ast.s_and_or.right, env));
-	}
+		return (process_binary_heredoc(node->u_ast.s_and_or.left,
+				node->u_ast.s_and_or.right, env));
 	if (node->type == AST_SUBSHELL)
 	{
 		if (process_heredocs(node->u_ast.s_subshell.and_or, env) == -1)
@@ -75,142 +76,4 @@ int	process_heredocs(t_ast *node, t_shell_env *env)
 	if (node->type == AST_SIMPLE_CMD)
 		return (process_simple_cmd_heredocs(node, env));
 	return (0);
-}
-
-static void	unlink_heredoc_node(t_ast *node)
-{
-	if (!node)
-		return ;
-	if (node->type == AST_IO_FILE)
-	{
-		if (node->u_ast.s_io_file.op->type == DLESS
-			|| node->u_ast.s_io_file.op->type == LESS)
-		{
-			if (node->u_ast.s_io_file.filename
-				&& ft_strncmp(node->u_ast.s_io_file.filename, "/tmp/.heredoc_",
-					14) == 0)
-			{
-				unlink(node->u_ast.s_io_file.filename);
-				free((char *)node->u_ast.s_io_file.filename);
-				node->u_ast.s_io_file.filename = NULL;
-			}
-		}
-	}
-}
-
-static void	free_heredoc_mem(t_ast *node)
-{
-	if (!node)
-		return ;
-	if (node->type == AST_IO_FILE)
-	{
-		if (node->u_ast.s_io_file.filename
-			&& ft_strncmp(node->u_ast.s_io_file.filename, "/tmp/.heredoc_",
-				14) == 0)
-		{
-			free((char *)node->u_ast.s_io_file.filename);
-			node->u_ast.s_io_file.filename = NULL;
-		}
-	}
-}
-
-static void	cleanup_simple_cmd_heredocs(t_ast *ast)
-{
-	t_ast	*curr;
-
-	curr = ast->u_ast.s_simple_cmd.cmd_prefix;
-	while (curr)
-	{
-		unlink_heredoc_node(curr->u_ast.s_cmd_prefix.io_file);
-		curr = curr->u_ast.s_cmd_prefix.cmd_prefix;
-	}
-	curr = ast->u_ast.s_simple_cmd.cmd_suffix;
-	while (curr)
-	{
-		unlink_heredoc_node(curr->u_ast.s_cmd_suffix.io_file);
-		curr = curr->u_ast.s_cmd_suffix.cmd_suffix;
-	}
-}
-
-void	cleanup_heredoc_files(t_ast *ast)
-{
-	if (!ast)
-		return ;
-	if (ast->type == AST_PIPE_SEQ)
-	{
-		cleanup_heredoc_files(ast->u_ast.s_pipe_seq.left);
-		cleanup_heredoc_files(ast->u_ast.s_pipe_seq.right);
-	}
-	else if (ast->type == AST_AND_OR || ast->type == AST_LIST)
-	{
-		cleanup_heredoc_files(ast->u_ast.s_and_or.left);
-		cleanup_heredoc_files(ast->u_ast.s_and_or.right);
-	}
-	else if (ast->type == AST_SUBSHELL)
-	{
-		cleanup_heredoc_files(ast->u_ast.s_subshell.and_or);
-		unlink_heredoc_node(ast->u_ast.s_subshell.io_file);
-	}
-	else if (ast->type == AST_CMD_PREFIX)
-	{
-		unlink_heredoc_node(ast->u_ast.s_cmd_prefix.io_file);
-		cleanup_heredoc_files(ast->u_ast.s_cmd_prefix.cmd_prefix);
-	}
-	else if (ast->type == AST_SIMPLE_CMD)
-		cleanup_simple_cmd_heredocs(ast);
-}
-
-static void	free_simple_cmd_heredocs(t_ast *ast)
-{
-	t_ast	*curr;
-
-	curr = ast->u_ast.s_simple_cmd.cmd_prefix;
-	while (curr)
-	{
-		free_heredoc_mem(curr->u_ast.s_cmd_prefix.io_file);
-		curr = curr->u_ast.s_cmd_prefix.cmd_prefix;
-	}
-	curr = ast->u_ast.s_simple_cmd.cmd_suffix;
-	while (curr)
-	{
-		free_heredoc_mem(curr->u_ast.s_cmd_suffix.io_file);
-		curr = curr->u_ast.s_cmd_suffix.cmd_suffix;
-	}
-}
-
-static void	free_heredoc_prefix(t_ast *prefix)
-{
-	t_ast	*curr;
-
-	curr = prefix;
-	while (curr)
-	{
-		free_heredoc_mem(curr->u_ast.s_cmd_prefix.io_file);
-		curr = curr->u_ast.s_cmd_prefix.cmd_prefix;
-	}
-}
-
-void	free_heredoc_filenames(t_ast *ast)
-{
-	if (!ast)
-		return ;
-	if (ast->type == AST_PIPE_SEQ)
-	{
-		free_heredoc_filenames(ast->u_ast.s_pipe_seq.left);
-		free_heredoc_filenames(ast->u_ast.s_pipe_seq.right);
-	}
-	else if (ast->type == AST_AND_OR || ast->type == AST_LIST)
-	{
-		free_heredoc_filenames(ast->u_ast.s_and_or.left);
-		free_heredoc_filenames(ast->u_ast.s_and_or.right);
-	}
-	else if (ast->type == AST_SUBSHELL)
-	{
-		free_heredoc_filenames(ast->u_ast.s_subshell.and_or);
-		free_heredoc_mem(ast->u_ast.s_subshell.io_file);
-	}
-	else if (ast->type == AST_CMD_PREFIX)
-		free_heredoc_prefix(ast);
-	else if (ast->type == AST_SIMPLE_CMD)
-		free_simple_cmd_heredocs(ast);
 }
